alert-service/requirements.txt
alembic
click==7.1.2
Flask==1.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
flask_cors
flask-swagger-ui==3.36.0
celery==5.1.2
redis==3.5.3
amqp==5.0.6
apscheduler==3.8.0 
flask-apscheduler==1.12.2 
python-dateutil==2.8.2 
tzlocal==2.1
requests
python-socketio[client]
gunicorn